requirements.txt
# testing
pytest==2.6.4
pytest-django==2.8.0
pytest-cov==1.8.1
mock==1.3.0
requests==2.7.0
responses==0.5.0
